requirements.txt
torch
torchvision
numpy
pillow
tqdm
scikit-learn
matplotlib
pyyaml